+++ examples/authentication/auth/config.js
@@ -1,0 +1,5 @@
+module.exports = {
+    JWT_SECRET: '__SECRET__',
+    COOKIE_NAME: 'auth_token',
+    AUTH_TOKEN_TTL: 60*60*24*7
+};