+++ examples/authentication/auth/context.js
@@ -1,0 +1,5 @@
+module.exports = (req = null) => {
+    return {
+        user: req?.user ?? 'system'
+    };
+};