+++ examples/authentication/readModels/user.resolvers.js
@@ -1,0 +1,9 @@
+const tableName = 'Users';
+
+module.exports = {
+    get: async (store, args, context) => {
+        if(context.user !== 'system') // users can only query their own entry
+            args.id = context.user.id;
+        return await store.findOne(tableName, args);
+    }
+};