+++ src/core/Blackrik.js
@@ -11,7 +11,7 @@ const CommandHandler = require('./CommandHandler');
 const QueryHandler = require('./QueryHandler');
 const CommandScheduler = require('./CommandScheduler');
 
-const Workflow = require('./Workflow');
+const Workflow = require('./workflow');
 
 const httpMethods = require('../resources/httpMethods');
 
@@ -128,16 +128,14 @@ class Blackrik
             )
         );
     }
-    
+
     _getSideEffectsProxy(sideEffects, event, config)
     {
         const blackrik = this;
         return new Proxy(sideEffects, {
             get: (target, prop, ...rest) => {
                 if(event.isReplay && config.noopSideEffectsOnReplay !== false)
-                {
                     return () => {};
-                }
                 
                 for(const [fn, argCount] of Object.entries({
                     'executeCommand': 1,
@@ -154,6 +152,7 @@ class Blackrik
                         return this[fn](...params);
                     }.bind(blackrik);
                 }
+
                 return Reflect.get(target, prop, ...rest);
             }
         });
@@ -202,13 +201,42 @@ class Blackrik
 
     _registerInternalMiddlewares()
     {
-        this.#server.use((...[req, , next]) => (req.blackrik = Object.freeze(this.#instance)) && next());
+        const instance = Object.freeze(this.#instance);
+        this.#server.use((...[req, , next]) => (req.blackrik = instance) && next());
     }
 
     _registerMiddlewares()
     {
         const {middlewares} = this.config.server;
-        middlewares.forEach(middleware => this.#server.use(...(Array.isArray(middleware) ? middleware : [middleware])));
+        // middlewares.forEach(middleware => this.#server.use(...(Array.isArray(middleware) ? middleware : [middleware]))); // Express 5 (or higher)
+        // Express < 5
+        middlewares.forEach(middleware => {
+            if(Array.isArray(middleware))
+            {
+                const [path, ...callbacks] = middleware;
+                middleware = [path, ...callbacks.map(callback => RequestHandler.catch(callback))];
+            }
+            else 
+                middleware = [RequestHandler.catch(middleware)];
+            this.#server.use(...middleware);
+        });
+    }
+
+    _registerErrorHandlingMiddlewares()
+    {
+        // http://expressjs.com/en/guide/error-handling.html
+        // Asynchronous route handlers, middleware must call next(err) otherwise its an unhandled error
+        // Starting with Express 5 route handlers and middleware that return a Promise will call next(value) automatically when they reject or throw an error
+
+        //TODO use this when Express 5 is production ready
+        // this.#server.use((e, req, res, next) => {
+        //     if(!e.status)
+        //     {
+        //         console.error(e);
+        //         return res.sendStatus(500).end(); // do not expose critical errors
+        //     }
+        //     return res.status(e.status).send(e.message || e).end();
+        // });
     }
 
     _processMiddlewares()
@@ -234,7 +262,8 @@ class Blackrik
             if(!path.startsWith('/'))
                 path = '/' + path;
 
-            this.#server.route(path)[method](callback);
+            // this.#server.route(path)[method](callback);
+            this.#server.route(path)[method](RequestHandler.catch(callback));
         });
     }
 
@@ -242,6 +271,13 @@ class Blackrik
     {
         this._registerInternalAPI();
         this._registerAPI();
+    }
+
+    buildContext(req = null)
+    {
+        return {
+            ...(this.config?.contextProvider(req) ?? {})
+        };
     }
 
     async start()
@@ -273,6 +309,7 @@ class Blackrik
         this.#server = new Server(this.config.server.config);
         this._processMiddlewares();
         this._processAPI();
+        this._registerErrorHandlingMiddlewares();
         await this.#server.start();
 
         return this;
@@ -280,7 +317,7 @@ class Blackrik
 
     async executeCommand(command, causationEvent = null)
     {
-        const context = {blackrik: this.#instance};
+        const context = this.buildContext();
         if(causationEvent)
             context.causationEvent = causationEvent;
         return !!await this._commandHandler.process(
@@ -303,7 +340,8 @@ class Blackrik
         return await this._queryHandler.process(
             readModel,
             resolver,
-            query
+            query,
+            this.buildContext()
         );
     }
 }