+++ src/core/Errors.js
@@ -1,9 +1,9 @@
 class BaseError extends Error
 {
-    constructor(message = 'Internal Error', code = 500)
+    constructor(message = 'Internal Error', status = 500)
     {
         super(message);
-        this.code = code;
+        this.status = status;
     }
 }
 module.exports.BaseError = BaseError;
@@ -13,7 +13,7 @@ class BadRequestError extends BaseError
     constructor(message = 'Bad Request')
     {
         super(message);
-        this.code = 400;
+        this.status = 400;
     }
 }
 module.exports.BadRequestError = BadRequestError;
@@ -23,7 +23,7 @@ class UnauthorizedError extends BaseError
     constructor(message = 'Unauthorized')
     {
         super(message);
-        this.code = 401;
+        this.status = 401;
     }
 }
 module.exports.UnauthorizedError = UnauthorizedError;
@@ -33,7 +33,7 @@ class ForbiddenError extends BaseError
     constructor(message = 'Forbidden')
     {
         super(message);
-        this.code = 403;
+        this.status = 403;
     }
 }
 module.exports.ForbiddenError = ForbiddenError;
@@ -43,7 +43,7 @@ class NotFoundError extends BaseError
     constructor(message = 'Not Found')
     {
         super(message);
-        this.code = 404;
+        this.status = 404;
     }
 }
 module.exports.NotFoundError = NotFoundError;
@@ -53,7 +53,7 @@ class ConflictError extends BaseError
     constructor(message = 'Conflict')
     {
         super(message);
-        this.code = 409;
+        this.status = 409;
     }
 }
 module.exports.ConflictError = ConflictError;