+++ test/core/RequestHandler.test.js
@@ -26,7 +26,7 @@ describe('RequestHandler', () => {
     class Res
     {
         closed = false;
-        code;
+        statusCode;
         data;
 
         _checkConnection()
@@ -52,18 +52,18 @@ describe('RequestHandler', () => {
             return this.send(JSON.stringify(data));
         }
 
-        status(code)
+        status(status)
         {
             this._checkConnection();
-            this.code = code;
+            this.statusCode = status;
             return this;
         }
         
-        sendStatus(code)
+        sendStatus(status)
         {
-            if(code === 500) // no need to mock other codes here
+            if(status === 500) // no need to mock other status codes here
                 this.send('Internal Error');
-            this.status(code).end();
+            this.status(status).end();
             return this;
         }
     }
@@ -85,10 +85,11 @@ describe('RequestHandler', () => {
         const handler = jest.fn(() => {});
 
         const req = createReq();
-        await (new RequestHandler(handler))(req, createRes());
+        const res = createRes();
+        await (new RequestHandler(handler))(req, res);
 
         expect(handler).toHaveBeenCalledTimes(1);
-        expect(handler).toHaveBeenNthCalledWith(1, req);
+        expect(handler).toHaveBeenNthCalledWith(1, req, res);
     });
 
     test('processes returned handler value correctly', async () => {
@@ -109,7 +110,7 @@ describe('RequestHandler', () => {
         const res = createRes();
         await (new RequestHandler(handler))(createReq(), res);
 
-        expect(res.code).toBe(error.code);
+        expect(res.statusCode).toBe(error.status);
         expect(res.data).toBe(error.message);
     });
 
@@ -120,7 +121,7 @@ describe('RequestHandler', () => {
         const res = createRes();
         await (new RequestHandler(handler))(createReq(), res);
 
-        expect(res.code).toBe(500);
+        expect(res.statusCode).toBe(500);
         expect(res.data).not.toBe(errorMessage);
     });
 
@@ -130,7 +131,7 @@ describe('RequestHandler', () => {
             constructor()
             {
                 this.msg = 'Error message';
-                this.code = 418;
+                this.status = 418;
             }
 
             toString()
@@ -144,7 +145,7 @@ describe('RequestHandler', () => {
         const res = createRes();
         await (new RequestHandler(handler))(createReq(), res);
 
-        expect(res.code).toBe(error.code);
+        expect(res.statusCode).toBe(error.status);
         expect(res.data).toBe(error.msg);
     });
 });